Object acces check

+++ doclib-web/src/main/java/com/strelnikov/doclib/web/controllers/CatalogRestController.java
@@ -1,14 +1,21 @@
 package com.strelnikov.doclib.web.controllers;
 
 import com.strelnikov.doclib.dto.CatalogDto;
+import com.strelnikov.doclib.dto.PermissionDto;
+import com.strelnikov.doclib.model.roles.PermissionType;
 import com.strelnikov.doclib.service.CatalogActions;
+import com.strelnikov.doclib.service.SecurityActions;
 import com.strelnikov.doclib.service.exceptions.CannotDeleteMainCatalogException;
 import com.strelnikov.doclib.service.exceptions.UnitIsAlreadyExistException;
 import com.strelnikov.doclib.service.exceptions.UnitNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.annotation.Secured;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 
 @RestController
@@ -19,37 +26,83 @@ public class CatalogRestController {
     @Autowired
     private CatalogActions catalogAct;
 
+    @Autowired
+    private SecurityActions securityActions;
+
     @GetMapping("get/{id}")
     public ResponseEntity<CatalogDto> getCatalog(@PathVariable int id){
         try {
             CatalogDto catalogDto = catalogAct.loadCatalog(id);
-            return ResponseEntity.ok(catalogDto);
+            Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+            if (securityActions.checkPermission(catalogDto,auth.getName(), PermissionType.READING)) {
+                catalogDto = catalogAct.filterContentList(catalogDto,auth.getName(),PermissionType.READING);
+                return ResponseEntity.ok(catalogDto);
+            }else{
+                return ResponseEntity.status(403).build();
+            }
         }catch (UnitNotFoundException e){
             return ResponseEntity.notFound().build();
         }
     }
 
+    @GetMapping("get/{id}/permissions/get")
+    public ResponseEntity<List<PermissionDto>> getPermissionList(@PathVariable int id){
+        try {
+            CatalogDto catalogDto = catalogAct.loadCatalog(id);
+            Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+            if (securityActions.checkPermission(catalogDto,auth.getName(),PermissionType.READING)) {
+                return ResponseEntity.ok(securityActions.getObjectPermissions(catalogDto));
+            }else{
+                return ResponseEntity.status(403).build();
+            }
+        } catch (UnitNotFoundException e) {
+            return ResponseEntity.status(404).build();
+        }
+    }
+
+    @PostMapping("get/{id}/permissions/post")
+    @Secured({"ROLE_ADMIN"})
+    public ResponseEntity<List<PermissionDto>> postPermissionList(@RequestBody List<PermissionDto> permissionDtoList, @PathVariable int id){
+        try {
+            CatalogDto catalogDto = catalogAct.loadCatalog(id);
+            securityActions.updatePermissions(permissionDtoList);
+            return ResponseEntity.ok(securityActions.getObjectPermissions(catalogDto));
+        } catch (UnitNotFoundException e) {
+            return ResponseEntity.notFound().build();
+        }
+    }
+
     @PostMapping("post")
-    public ResponseEntity<Object> postCatalog(@RequestBody CatalogDto catalogDto){
+    public ResponseEntity<CatalogDto> postCatalog(@RequestBody CatalogDto catalogDto){
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
         try{
-            catalogDto = catalogAct.saveCatalog(catalogDto);
-            return ResponseEntity.ok(catalogDto);
+            if(securityActions.checkPermission(catalogAct.loadCatalog(catalogDto.getParentId()),auth.getName(),PermissionType.WRITING)){
+                catalogDto = catalogAct.saveCatalog(catalogDto);
+                return ResponseEntity.ok(catalogDto);
+            }
+            return ResponseEntity.status(403).build();
         }catch (UnitIsAlreadyExistException e) {
-            return ResponseEntity.badRequest().body("Catalog is already exist");
+            return ResponseEntity.badRequest().build();
+        } catch (UnitNotFoundException e) {
+            return ResponseEntity.notFound().build();
         }
     }
 
     @DeleteMapping("delete/{id}")
-    public ResponseEntity<Object> deleteCatalog(@PathVariable int id){
+    public ResponseEntity<CatalogDto> deleteCatalog(@PathVariable int id){
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
         try {
             CatalogDto catalogDto = catalogAct.loadCatalog(id);
-            catalogAct.deleteCatalog(catalogDto);
-            catalogDto = catalogAct.loadCatalog(catalogDto.getParentId());
-            return ResponseEntity.ok(catalogDto);
+            if(securityActions.checkPermission(catalogDto,auth.getName(),PermissionType.WRITING)){
+                catalogAct.deleteCatalog(catalogDto);
+                catalogDto = catalogAct.loadCatalog(catalogDto.getParentId());
+                return ResponseEntity.ok(catalogDto);
+            }
+            return ResponseEntity.status(403).build();
         } catch (UnitNotFoundException e) {
-            return ResponseEntity.badRequest().body("There is no catalog with id = " + id);
+            return ResponseEntity.badRequest().build();
         }catch (CannotDeleteMainCatalogException e){
-            return ResponseEntity.badRequest().body("Can't delete main catalog");
+            return ResponseEntity.badRequest().build();
         }
     }
 }
